+++ app/src/main/java/com/example/hangsha_android/ui/view/bookmarks/BookmarksViewModel.kt
@@ -240,11 +240,11 @@ private fun EventSummaryResponse.toBookmarkedEventItem(): BookmarkedEventItem {
     val dDayLabel = applyEndDate?.let { targetDate ->
         val diff = targetDate.toEpochDay() - LocalDate.now().toEpochDay()
         when {
-            diff == 0L -> "지원 D-day"
-            diff > 0L -> "지원 D-$diff"
-            else -> "지원 D$diff"
+            diff == 0L -> "D-day"
+            diff > 0L -> "D-$diff"
+            else -> "D$diff"
         }
-    } ?: "지원 -"
+    } ?: "-"
 
     return BookmarkedEventItem(
         id = id,
