+++ app/src/main/java/com/example/hangsha_android/ui/view/dailyevents/DailyEventsUiState.kt
@@ -42,6 +42,7 @@ data class DailyEventItem(
     val imageUrl: String?,
     val organization: String?,
     val eventEndDisplay: String,
+    val applyPeriodDisplay: String,
     val dDayLabel: String,
     val accentColor: Color,
     val isBookmarked: Boolean,
