+++ app/src/main/java/com/example/hangsha_android/ui/view/eventdetail/EventDetailUiState.kt
@@ -30,6 +30,7 @@ data class EventDetailItem(
     val organization: String?,
     val location: String?,
     val eventEndDisplay: String,
+    val applyPeriodDisplay: String,
     val dDayLabel: String,
     val eventTypeLabel: String,
     val eventTypeColor: Color,
